+++ config/sikessem-ui.php
@@ -1,5 +1,16 @@
 <?php
 
 return [
-
+    'flash' => [
+        'element' => 'p',
+        'attributes' => [],
+    ],
+    'flashes' => [
+        'element' => 'ul',
+        'attributes' => [],
+        'flash' => [
+            'element' => 'li',
+            'attributes' => [],
+        ],
+    ],
 ];