+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [Sikessem's UI kit](https://github.com/sikessem/ui) will 
 
 **Full Changelog:** [v0.0.0...v0.1.0](https://github.com/sikessem/ui/compare/v0.0.0...v0.1.0)
 
+- [0.x] Add [**flashes** component](https://github.com/sikessem/ui#flashes-component)
+- [0.x] Add [**flash** component](https://github.com/sikessem/ui#flash-component)
 - [0.x] Add [**label** component](https://github.com/sikessem/ui#label-component)
 - [0.x] Add [**menu** component](https://github.com/sikessem/ui#menu-component)
 - [0.x] Add [**button** component](https://github.com/sikessem/ui#button-component)